requirements.txt
fastapi
uvicorn
uvicorn[standard]
python-dotenv
motor
pydantic
pyjwt
httpx
pymongo
python-jose